fix(cli): exact seed flags via RETURNING, shared resolveDbUrl, deployment-mode guard
- ensureInstanceAdmin now derives createdUser/createdRole from RETURNING on
  the ON CONFLICT DO NOTHING inserts instead of a check-then-insert, so the
  flags (and the seeded/already-present message) are exact even when
  concurrent seeds race; the race test now asserts exactly one winner per row
- extract resolveDbUrl to cli/src/config/db.ts and reuse it from both
  auth-bootstrap-ceo and auth-seed-instance-admin
- mirror bootstrap-ceo's deployment-mode guard: exit early with an info
  message on local_trusted configs (skipped when no config file is readable,
  keeping headless --db-url/DATABASE_URL automation working)
