+++ cli/src/commands/auth-bootstrap-ceo.ts
@@ -4,6 +4,7 @@ import pc from "picocolors";
 import { and, eq, gt, isNull } from "drizzle-orm";
 import { createDb, instanceUserRoles, invites } from "@paperclipai/db";
 import { inferBindModeFromHost } from "@paperclipai/shared";
+import { resolveDbUrl } from "../config/db.js";
 import { loadPaperclipEnvFile } from "../config/env.js";
 import { readConfig, resolveConfigPath } from "../config/store.js";
 
@@ -13,20 +14,6 @@ function hashToken(token: string) {
 
 function createInviteToken() {
   return `pcp_bootstrap_${randomBytes(24).toString("hex")}`;
-}
-
-function resolveDbUrl(configPath?: string, explicitDbUrl?: string) {
-  if (explicitDbUrl) return explicitDbUrl;
-  const config = readConfig(configPath);
-  if (process.env.DATABASE_URL) return process.env.DATABASE_URL;
-  if (config?.database.mode === "postgres" && config.database.connectionString) {
-    return config.database.connectionString;
-  }
-  if (config?.database.mode === "embedded-postgres") {
-    const port = config.database.embeddedPostgresPort ?? 54329;
-    return `postgres://paperclip:paperclip@127.0.0.1:${port}/paperclip`;
-  }
-  return null;
 }
 
 function resolveBaseUrl(configPath?: string, explicitBaseUrl?: string) {
